use exit status instead of STDOUT
Version 6.8.8-9 of ImageMagick's compare returns 2 on error, 0 if the images are similar, and 1 if they are dissimilar. Rather than using the output of compare, this changes the code to use the actual return code.

When using compare, the actual output of the command is "0 @ 0,0" if the images are similar. I'm presuming, based on the previous code, that "0" was output to STDOUT (or STDERR) if the images were similar.